[FIX] minor warning
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.10@10623 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/VISUAL/MultiGradient.h
@@ -146,7 +146,7 @@ public:
     {
       OPENMS_PRECONDITION(pre_.size() != 0, "MultiGradient::precalculatedColorByIndex(Int): Precalculation mode not activated!");
       OPENMS_PRECONDITION( index >= 0, "MultiGradient::precalculatedColorByIndex(Int): negative indexes not allowed");
-      OPENMS_PRECONDITION( index < pre_.size(), (String("MultiGradient::indexedColor(Int): index ") + index + " out of specified range (0-" + pre_.size() + ")!").c_str());
+      OPENMS_PRECONDITION( index < (Int)pre_.size(), (String("MultiGradient::indexedColor(Int): index ") + index + " out of specified range (0-" + pre_.size() + ")!").c_str());
 
       return pre_[index];
     }
